feat: add partial download (per model) mode
* feat: add keras additive and multiplicative attention layers

* Download feature refactoring: backup commit

* Backup config for download

* Minor bug fixes in download

* Minor fixes in download.py

* Fixed download() in test_quick_start.py

* Added downloads to ranking model config metadata

* Updated configs for gobot_dstc2, odqa

* Downloads fixing backup commit

* feat: add download metadata for the error_model

* feat: add download links to configs

* Refactored download procedure

* fixed dataset subdir in gobot_dstc2.json

* feat: add metadata for go_bots

* Added -d/--download argument for al deeppavlov.deep.main()

* feat: add url metadata for seq2seq_go_bot

* added recurcive search for componets in config references

* feat: add download to squad configs

* Re-fixed tests/test_quick_start.py

* Fixed deeppavlov/configs/go_bot/gobot_dstc2_best.json

* feat: add vocabs dependency to seq2seq_go_bot

* Small fixes in core/data/utils.py

* Fixed error model: karta slov

* fixed ranking

* fix: add vocabs for download

* feat: urls for NER added to configs

* fix:  typo fix in fonfig

* feat: rm extra downloads for gobot_dstc2

* feat: add vocab and embeddings download for configs without pretrained models

* fix: ner download files added to config

* feat: add download in config

* added deep.py download mode, fixed recursive dependencies downloads

* fixed deeppavlov/configs/go_bot/gobot_dstc2_all.json

* fix: ner download urls fixed

* Minor fixes in intents configs

* Fixed deeppavlov/core/data/utils.py

* small fixes in deeppavlov/core/data/utils.py

* feat: show detailed output when tests break in train

* fixed download.py for compatibility with test mode

* fixed error_model configs

* restored error_model configs

* Fixed test errors in error_model and intents

* fixed error_model_custom_vocab

* fix: update all ODQA configs metadata

+++ deeppavlov/core/data/utils.py
@@ -22,6 +22,7 @@ import tarfile
 import gzip
 import re
 import zipfile
+import shutil
 
 from deeppavlov.core.common.log import get_logger
 
@@ -34,25 +35,33 @@ tqdm.monitor_interval = 0
 
 
 def download(dest_file_path, source_url, force_download=True):
-    """Download a file from URL
+    """Download a file from URL to one or several target locations
 
     Args:
-        dest_file_path: path to the file destination file (including file name)
+        dest_file_path: path or list of paths to the file destination files (including file name)
         source_url: the source URL
         force_download: download file if it already exists, or not
 
     """
     CHUNK = 16 * 1024
-    dest_file_path = Path(dest_file_path).absolute()
 
-    if force_download or not dest_file_path.exists():
-        dest_file_path.parent.mkdir(parents=True, exist_ok=True)
+    if isinstance(dest_file_path, str):
+        dest_file_path = [Path(dest_file_path).absolute()]
+    elif isinstance(dest_file_path, Path):
+        dest_file_path = [dest_file_path.absolute()]
+    elif isinstance(dest_file_path, list):
+        dest_file_path = [Path(path) for path in dest_file_path]
+
+    first_dest_path = dest_file_path.pop()
+
+    if force_download or not first_dest_path.exists():
+        first_dest_path.parent.mkdir(parents=True, exist_ok=True)
 
         r = requests.get(source_url, stream=True)
         total_length = int(r.headers.get('content-length', 0))
 
-        with dest_file_path.open('wb') as f:
-            log.info('Downloading from {} to {}'.format(source_url, dest_file_path))
+        with first_dest_path.open('wb') as f:
+            log.info('Downloading from {} to {}'.format(source_url, first_dest_path))
 
             pbar = tqdm(total=total_length, unit='B', unit_scale=True)
             for chunk in r.iter_content(chunk_size=CHUNK):
@@ -61,7 +70,16 @@ def download(dest_file_path, source_url, force_download=True):
                     f.write(chunk)
             f.close()
     else:
-        log.info('File already exists in {}'.format(dest_file_path))
+        log.info('File already exists in {}'.format(first_dest_path))
+
+    while len(dest_file_path) > 0:
+        dest_path = dest_file_path.pop()
+
+        if force_download or not dest_path.exists():
+            dest_path.parent.mkdir(parents=True, exist_ok=True)
+            shutil.copy(str(first_dest_path), str(dest_path))
+        else:
+            log.info('File already exists in {}'.format(dest_path))
 
 
 def untar(file_path, extract_folder=None):
@@ -103,32 +121,46 @@ def ungzip(file_path, extract_folder=None):
             fout.write(block)
 
 
-def download_decompress(url, download_path, extract_path=None):
-    """Download and extract .tar.gz or .gz file. The archive is deleted after extraction.
+def download_decompress(url, download_path, extract_paths=None):
+    """Download and extract .tar.gz or .gz file to one or several target locations.
+    The archive is deleted if extraction was successful.
 
     Args:
         url: URL for file downloading
         download_path: path to the directory where downloaded file will be stored
         until the end of extraction
-        extract_path: path where contents of archive will be extracted
+        extract_paths: path or list of paths where contents of archive will be extracted
     """
     file_name = url.split('/')[-1]
     download_path = Path(download_path)
-    if extract_path is None:
-        extract_path = download_path
-    extract_path = Path(extract_path)
     arch_file_path = download_path / file_name
-    log.info('Extracting {} archive into {}'.format(arch_file_path, extract_path))
     download(arch_file_path, url)
-    if url.endswith('.tar.gz'):
-        untar(arch_file_path, extract_path)
-    elif url.endswith('.gz'):
-        ungzip(arch_file_path, extract_path)
-    elif url.endswith('.zip'):
-        zip_ref = zipfile.ZipFile(arch_file_path, 'r')
-        zip_ref.extractall(extract_path)
-        zip_ref.close()
-    arch_file_path.unlink()
+
+    if extract_paths is None:
+        extract_paths = [download_path]
+    elif isinstance(extract_paths, str):
+        extract_paths = [Path(extract_paths)]
+    elif isinstance(extract_paths, list):
+        extract_paths = [Path(path) for path in extract_paths]
+
+    if url.endswith(('.tar.gz', '.gz', '.zip')):
+        for extract_path in extract_paths:
+            log.info('Extracting {} archive into {}'.format(arch_file_path, extract_path))
+            extract_path.mkdir(parents=True, exist_ok=True)
+
+            if url.endswith('.tar.gz'):
+                untar(arch_file_path, extract_path)
+            elif url.endswith('.gz'):
+                ungzip(arch_file_path, extract_path)
+            elif url.endswith('.zip'):
+                zip_ref = zipfile.ZipFile(arch_file_path, 'r')
+                zip_ref.extractall(extract_path)
+                zip_ref.close()
+
+        arch_file_path.unlink()
+    else:
+        log.error('File {} has unsupported format. '
+                  'Not extracted, downloaded to {}'.format(file_name, arch_file_path))
 
 
 def load_vocab(vocab_path):
